Separate tests into feature specific classes.
