+++ tests/Feature/Admin/UserAdministrationTest.php
@@ -45,80 +45,4 @@ class UserAdministrationTest extends TestCase
             ->get(route('admin.users.index'))
             ->assertViewHas('users', $users->fresh());
     }
-
-    /** @test */
-    public function an_administrator_can_manually_confirm_a_users_email_address()
-    {
-        $user = create(User::class, [
-            'confirmed' => false
-        ]);
-
-        $this->assertFalse($user->confirmed);
-
-        $this->signInAdmin()
-            ->patch(route('admin.confirm-user.update', ['id' => $user->id]))
-            ->assertRedirect(route('admin.users.index'));
-
-        $this->assertTrue($user->fresh()->confirmed);
-    }
-
-    /** @test */
-    public function non_administrators_cannot_confirm_a_users_email_address()
-    {
-        $regularUser = create(User::class);
-
-        $this->actingAs($regularUser)
-            ->get(route('admin.users.index'))
-            ->assertStatus(Response::HTTP_FORBIDDEN);
-    }
-
-    // User Suspension Tests
-
-    /** @test */
-    public function administrators_can_suspend_a_users_account()
-    {
-        $user = create(User::class, ['active' => true]);
-
-        $this->assertTrue($user->active);
-
-        $this->signInAdmin()
-            ->delete(route('admin.suspend-user.destroy', $user))
-            ->assertRedirect(route('admin.users.index'));
-
-        $this->assertFalse($user->fresh()->active, 'Failed asserting that the user is suspended.');
-    }
-
-    /** @test */
-    public function non_administrators_cannot_suspend_a_users_account()
-    {
-        $user = create(User::class);
-
-        $this->signIn()
-            ->delete(route('admin.suspend-user.destroy', $user))
-            ->assertStatus(Response::HTTP_FORBIDDEN);
-    }
-
-    /** @test */
-    public function administrators_can_activate_a_users_account()
-    {
-        $user = create(User::class, ['active' => false]);
-
-        $this->assertFalse($user->active);
-
-        $this->signInAdmin()
-            ->post(route('admin.suspend-user.store', $user))
-            ->assertRedirect(route('admin.users.index'));
-
-        $this->assertTrue($user->fresh()->active, 'Failed asserting that the user is active.');
-    }
-
-    /** @test */
-    public function non_administrators_cannot_activate_a_users_account()
-    {
-        $user = create(User::class);
-
-        $this->signIn()
-            ->post(route('admin.suspend-user.store', $user))
-            ->assertStatus(Response::HTTP_FORBIDDEN);
-    }
 }
